fixed some types that were redundant
